refactor(#109): 좌석 생성시 공연장 좌석 정보와 일치 하는지 검증, 공연 목록 조회 pageable 검증 추가
